application.properties, user, type and net api doc
