+++ src/main/java/com/juansenen/gaticket/domain/Incidences.java
@@ -1,6 +1,7 @@
 package com.juansenen.gaticket.domain;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -17,23 +18,30 @@ public class Incidences {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "incidence_id")
+    @Schema(accessMode = Schema.AccessMode.READ_ONLY, description = "incidenceId", example = "12")
     private long incidencesId;
     @Column(name = "incidence_commit")
+    @Schema(description = "Description", example = "My computer has began to go slowly. And I can´t see mp4 videos")
     private String incidenceCommit;
     @Column(name = "incidence_status")
+    @Schema(description = "Incidence status", example = "true" )
     private boolean incidenceStatus;
     @Column(name = "incidence_date")
+    @Schema(description = "Incidence date start", example = "04/11/2023", format = "date")
     @JsonFormat(pattern = "MM/dd/yyyy")
     private Date incidenceDate;
     @Column(name="incidence_datefinish")
+    @Schema(description = "Incidence date fix it", example = "06/11/2023", format = "date")
     @JsonFormat(pattern = "MM/dd/yyyy")
     private Date incidenceDateFinish;
 
     @ManyToOne
     @JoinColumn(name = "device")
+    @Schema(description = "The associated device")
     private Device device;
 
     @ManyToOne
     @JoinColumn(name = "user")
+    @Schema(description = "The user who reported de incidence")
     private User user; //Clave de la tabla principal
 }
