+++ src/main/resources/application.properties
@@ -26,3 +26,5 @@ management.endpoint.health.show-details=always
 
 # Configuracion Swagger
 springdoc.swagger-ui.operationsSorter=method
+#Configuracion para que solo muestre los endpoints de los controller
+springdoc.packages-to-scan=com.juansenen.gaticket.controllers
